% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBoxTransform.R
\name{BoxCoxTrans}
\alias{BoxCoxTrans}
\title{Transform data with BoxCox Transformation}
\usage{
BoxCoxTrans(data, x, y)
}
\arguments{
\item{data}{A data frame}

\item{x}{The name of the x variable from the data frame}

\item{y}{The name of the y variable from the data frame}
}
\value{
A list containing the BoxCox transformed data
}
\description{
Transform data with BoxCox Transformation
}
\author{
Jessica Schaub
}
